ConfigProvider useConfig refactoring
