+++ src/components/Toast/Toast.tsx
@@ -5,7 +5,7 @@ import dayjs from 'dayjs';
 import { animated, useSpring } from '@react-spring/web';
 
 import { ToastDetail } from 'providers/ToastsProvider';
-import { useConfig } from 'providers/ConfigProvider';
+import { useConfig } from 'hooks/useConfig';
 
 import Icon from 'components/Icon/Icon';
 import Button from 'components/Button/Button';
