+++ src/pages/UsersPage/api/useGetUsers.ts
@@ -2,7 +2,7 @@ import { UseQueryResult, useQuery } from '@tanstack/react-query';
 import { User } from 'api/useGetUser';
 
 import { useAxios } from 'hooks/useAxios';
-import { useConfig } from 'providers/ConfigProvider';
+import { useConfig } from 'hooks/useConfig';
 import { QueryKeys } from 'utils/constants';
 
 /**
