+++ src/providers/__tests__/ConfigProvider.test.tsx
@@ -1,9 +1,8 @@
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import { renderHook as renderHookWithoutWrapper } from '@testing-library/react';
 
-import { render, renderHook, screen, waitFor } from 'test/test-utils';
+import { render, screen } from 'test/test-utils';
 
-import ConfigContextProvider, { useConfig } from 'providers/ConfigProvider';
+import ConfigContextProvider from 'providers/ConfigProvider';
 
 describe('ConfigProvider', () => {
   it('should render successfully', async () => {
@@ -43,22 +42,5 @@ describe.skip('ConfigProvider error', () => {
 
     // ASSERT
     expect(renderContextProvider).toThrow(/is a required field/);
-  });
-});
-
-describe('useConfig', () => {
-  it('should return the context', async () => {
-    // ARRANGE
-    const { result } = renderHook(() => useConfig());
-    await waitFor(() => expect(result.current).not.toBeNull());
-
-    // ASSERT
-    expect(result.current).toBeDefined();
-    expect(result.current.VITE_BUILD_ENV_CODE).toBe('test');
-  });
-
-  it('should throw error when not within provider', () => {
-    // ASSERT
-    expect(() => renderHookWithoutWrapper(() => useConfig())).toThrow(/hook must be used within/);
   });
 });
